PIMCORE-1908 - problem with PHP glob() function
git-svn-id: http://www.pimcore.org/svn/pimcore/private/core/trunk@2634 1f8fe7d8-47f0-464c-8d0a-336f4953ab05

+++ pimcore/models/Asset/Video.php
@@ -48,8 +48,11 @@ class Asset_Video extends Asset {
             $this->setCustomSetting("thumbnails", null);
 
             // video thumbnails and image previews
-            foreach (glob(PIMCORE_TEMPORARY_DIRECTORY . "/video_" . $this->getId() . "__*") as $file) {
-                unlink($file);
+            $files = glob(PIMCORE_TEMPORARY_DIRECTORY . "/video_" . $this->getId() . "__*");
+            if(is_array($files)) {
+                foreach ($files as $file) {
+                    unlink($file);
+                }
             }
         }
     }
